+++ web/geosearch/datapunt_geosearch/exceptions.py
@@ -1,0 +1,2 @@
+class DataSourceException(Exception):
+    pass
